requirements.txt
anthropic>=0.49.0
firecrawl-py>=1.5.0
apscheduler>=3.10.4
requests>=2.31.0
jinja2>=3.1.4
python-dotenv>=1.0.0
pyyaml>=6.0.1
